% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCMSAlignerModel-method.R
\name{alignToModel.nn}
\alias{alignToModel.nn}
\title{#' Insert a feature into the information table.
#'
#' @param lam An LCMS aligner model
#' @param peak A peak information
#' @param pos THe position in whichi to sinsert the data
#'
#' @return The LCMSAlignerModel
#' @export
#'
#' @examples
insertPointInFeature <- function(lam, peak, cor_peak, pos, id_sample) {
  lam@data[[pos]] <- extendDfRow(lam@data[[pos]], lam@peaks[pos, "current_num"])
  lam@data[[pos]][lam@peaks[pos, "current_num"], ] <- as.numeric(c(peak[[1]], peak[[2]],cor_peak[[2]],peak[[3]],id_sample))
  return(lam)
}
Align a peaktable to a model. The index are returned because the majroity of the data are not considered enough}
\usage{
alignToModel.nn(lam, peaktable, id_sample, threshold = 5)
}
\arguments{
\item{lam}{An LCMSAlignerModel}

\item{peaktable}{A peak table to be added to the model}

\item{id_sample}{The name of the ID to use eventually}

\item{threshold}{The threshold in diveergence (5 should be fine :) used ot threshold the model}
}
\value{
Am LCMSAlignerModel with the added peaktable.
}
\description{
#' Insert a feature into the information table.
#'
#' @param lam An LCMS aligner model
#' @param peak A peak information
#' @param pos THe position in whichi to sinsert the data
#'
#' @return The LCMSAlignerModel
#' @export
#'
#' @examples
insertPointInFeature <- function(lam, peak, cor_peak, pos, id_sample) {
  lam@data[[pos]] <- extendDfRow(lam@data[[pos]], lam@peaks[pos, "current_num"])
  lam@data[[pos]][lam@peaks[pos, "current_num"], ] <- as.numeric(c(peak[[1]], peak[[2]],cor_peak[[2]],peak[[3]],id_sample))
  return(lam)
}
Align a peaktable to a model. The index are returned because the majroity of the data are not considered enough
}
