% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCMSAlignerModel-method.R
\name{buildDataMatrix,LCMSAlignerModel-method}
\alias{buildDataMatrix,LCMSAlignerModel-method}
\title{Title}
\usage{
\S4method{buildDataMatrix}{LCMSAlignerModel}(
  object,
  subsample = NULL,
  subvariable = NULL,
  quant_var = "int",
  add_summary = TRUE,
  summary_vars = c("peakwidth", "rt", "rt_cor", "right_on_left_assymetry")
)
}
\arguments{
\item{LCMSAlignerModel}{}
}
\value{

}
\description{
Title
}
\examples{
print("example to be put here")
}
