% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCMSAlignerModel-method.R
\name{exportDataMatrix}
\alias{exportDataMatrix}
\title{Title}
\usage{
exportDataMatrix(
  object,
  path,
  subsample = NULL,
  subvariable = NULL,
  quant_var = "int",
  add_summary = TRUE,
  summary_vars = c("peakwidth", "rt", "rt_cor", "right_on_left_assymetry"),
  max_memory = 2000
)
}
\arguments{
\item{LCMSAlignerModel}{}
}
\value{

}
\description{
Title
}
\examples{
print("example to be put here")
}
