% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCMSAlignerModel-method.R
\name{findCandidates}
\alias{findCandidates}
\title{Find cndidates to put into a peaktable.}
\usage{
findCandidates(lar, rtree, peak)
}
\arguments{
\item{lar}{The reference object of an LCMSAlignerModel}

\item{rtree}{The Rtree of LC-MS  map data.}

\item{peak}{The peak to be inserted.}
}
\value{
The indices of the candidates
}
\description{
Find cndidates to put into a peaktable.
}
\examples{
print("Examples to be put here")
}
