% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{LCMSAlignerModel-class}
\alias{LCMSAlignerModel-class}
\title{LCMSAlignerModel}
\description{
LCMSAlignerModel
}
\section{Slots}{

\describe{
\item{\code{references}}{The mains paramters which are reused.}

\item{\code{peaks}}{A dataframe storing informations about ALL the peaks at any stage.}

\item{\code{data}}{A list of all storing all the occurences of peaks currently in memoery}

\item{\code{files}}{LCMSAlignerFileHandler The file handler system ensuring that multiple files are not processed multiple files}

\item{\code{storage}}{LCMSAlignerDataStorage A file system storing all the data}

\item{\code{rt_correction}}{A list storing all the corrected RTs just for fun.}

\item{\code{files_in_memory}}{A list storing the ids of all the character in memory.}
}}

\examples{
print("Examples to be put here")
}
