% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCMSAlignerModel-initialisation.R
\name{LCMSAlignerModelFromDirectory}
\alias{LCMSAlignerModelFromDirectory}
\title{Align all the file form a directory to an exisiting LCMSAlignerModel eventually}
\usage{
LCMSAlignerModelFromDirectory(
  directory,
  path_model,
  output,
  save_interval = 5,
  num_file = 10,
  reset = FALSE,
  threshold = 5,
  span = 0.6,
  n_clusters = 15,
  supp_data = c("peakwidth", "SN", "right_on_left_assymetry"),
  ransac_l1 = 0.05,
  ransac_dist_threshold = 0.1,
  maxAlign = 100,
  seed = 512,
  max_cor = 0.1,
  clustering = TRUE,
  bpp = NULL,
  graphical = FALSE,
  ...
)
}
\arguments{
\item{directory}{The directory to monitor}

\item{path_model}{The path to store the model}

\item{output}{The directory of output for for the ocnstructed block file}

\item{save_interval}{The number of aligned peak tbale between save}

\item{num_file}{If A new modle is created, how many files are used to find the references peaks}

\item{reset}{Shall the processing be restarted from the beginning eventually}

\item{span}{Smoothing parameters contorlling the loess}

\item{n_clusters}{Numbers of clusters evnetually tested.}

\item{...}{Arguments which will be passed to LCMSAlignerModel}
}
\value{
The LCMSAlignerObjects
}
\description{
Align all the file form a directory to an exisiting LCMSAlignerModel eventually
}
\examples{
print("Examples to be put here")
}
