% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCMSAlignerReferences-method.R
\name{params,LCMSAlignerReferences-method}
\alias{params,LCMSAlignerReferences-method}
\title{Set the parameters on LCMSAlignReferences object}
\usage{
\S4method{params}{LCMSAlignerReferences}(lar)
}
\arguments{
\item{lar}{An LCMSAlignerReferencesobject}
}
\value{
A named list of the parameters of the dataset
}
\description{
Set the parameters on LCMSAlignReferences object
}
\examples{
print("Examples to be put here")
}
