% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCMSAlignerReferences-method.R
\docType{methods}
\name{params<-,LCMSAlignerReferences-method}
\alias{params<-,LCMSAlignerReferences-method}
\title{Set the parameters on LCMSAlignReferences object}
\usage{
\S4method{params}{LCMSAlignerReferences}(lar, values) <- value
}
\arguments{
\item{lar}{An LCMSAlignerReferencesobject}

\item{values}{A named list with names corresponding to the parameters.}
}
\value{
A filled LCMSAlignerReference sobject
}
\description{
Set the parameters on LCMSAlignReferences object
}
\examples{
print("Examples to be put here")
}
