% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCMSAlignerModel-method.R
\name{plotDevRt}
\alias{plotDevRt}
\title{The percentage of alignement in the data}
\usage{
plotDevRt(lam, int_threshold = 0.15)
}
\arguments{
\item{lam}{The LCMSalignerModel to align}

\item{int_threshold}{The alignement threshold to be used eventually.}
}
\value{
Nothing
}
\description{
The percentage of alignement in the data
}
\examples{
print("Examples to be put here")
}
