% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCMSAlignerModel-method.R
\name{readAlignment}
\alias{readAlignment}
\title{Reading an alignment from a file}
\usage{
readAlignment(path)
}
\arguments{
\item{path}{The path ofhte alignment object to read}
}
\value{
An LCMSAlignerModel object
}
\description{
Reading an alignment from a file
}
\examples{
print("Examples ot be put here")
}
