% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCMSAlignerModel-method.R
\name{writeAlignment}
\alias{writeAlignment}
\title{Writing an alignment to a file}
\usage{
writeAlignment(lam, path)
}
\arguments{
\item{lam}{An LCMSAlignerModel}

\item{path}{The path ofhte alignment object to read}
}
\value{
A boolean indicating if the file has been succesfully written
}
\description{
Writing an alignment to a file
}
\examples{
print("Examples ot be put here")
}
